fix(issue-167): align play_interactive.py with current env contract
- Extract RslRlVecEnvWrapper to shared module (unilab/utils/rsl_rl_vec_env_wrapper.py)
- Update play_interactive.py to use shared wrapper, removing duplicate implementation
- Update train_rsl_rl.py to use shared wrapper
- Remove old (_, obs, _) contract compatibility branch from reset()
- Add policy_obs_mode parameter support to shared wrapper
- Add tests for play_interactive.py wrapper contract behavior

Closes #167
