+++ src/unilab/utils/__init__.py
@@ -2,6 +2,7 @@
 from unilab.utils.algo_utils import build_actor, ensure_registries
 from unilab.utils.offpolicy_logger import OffPolicyLogger
 from unilab.utils.onpolicy_logger import OnPolicyLogger
+from unilab.utils.rsl_rl_vec_env_wrapper import RslRlVecEnvWrapper
 from unilab.utils.torch_utils import to_numpy, to_torch
 
 __all__ = [
@@ -11,4 +12,5 @@ __all__ = [
     "OnPolicyLogger",
     "ensure_registries",
     "build_actor",
+    "RslRlVecEnvWrapper",
 ]
